+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from './utils'
-export * from './types'
+export * from './utils/utils'
+export * from './utils/types'
 export * from './theme'
