+++ src/theme.ts
@@ -1,20 +1,20 @@
-import type { ColorType, rgbColor } from '.'
+import type { ColorType, RgbColor } from '.'
 import { convertColor, getColorType, isColor, rgbToHex, rgbToHsb, rgbToHsl } from './'
 
-function tint(components: rgbColor, intensity: number) {
-  return components.map(c => Math.round(c + (255 - c) * intensity)) as rgbColor
+function tint(components: RgbColor, intensity: number) {
+  return components.map(c => Math.round(c + (255 - c) * intensity)) as RgbColor
 }
 
-function shade(components: rgbColor, intensity: number) {
-  return components.map(c => Math.round(c * intensity)) as rgbColor
+function shade(components: RgbColor, intensity: number) {
+  return components.map(c => Math.round(c * intensity)) as RgbColor
 }
 
 function withTint(intensity: number) {
-  return (hex: rgbColor) => tint(hex, intensity)
+  return (hex: RgbColor) => tint(hex, intensity)
 }
 
 function withShade(intensity: number) {
-  return (hex: rgbColor) => shade(hex, intensity)
+  return (hex: RgbColor) => shade(hex, intensity)
 }
 
 export interface ThemeMetas {
@@ -85,7 +85,7 @@ export function theme(color: string, options: ThemeOptions = {}): ThemeMetas {
 
   const { type, render } = { ...defaultOptions, ...options } as Required<ThemeOptions>
 
-  const finnalRender = (meta: [keyof ThemeMetas, rgbColor]) => {
+  const finnalRender = (meta: [keyof ThemeMetas, RgbColor]) => {
     let cs = ''
 
     switch (type) {
@@ -114,7 +114,7 @@ export function theme(color: string, options: ThemeOptions = {}): ThemeMetas {
     200: withTint(0.75),
     300: withTint(0.6),
     400: withTint(0.3),
-    500: (c: rgbColor) => c,
+    500: (c: RgbColor) => c,
     600: withShade(0.9),
     700: withShade(0.6),
     800: withShade(0.45),
