+++ src/utils/types.ts
@@ -1,0 +1,48 @@
+/**
+ * [0-360, 0-100, 0-100]
+ */
+export type HslColor = [number, number, number]
+/**
+ * [0-360, 0-100, 0-100]
+ */
+export type HsbColor = [number, number, number]
+/**
+ * [0-255, 0-255, 0-255]
+ */
+export type RgbColor = [number, number, number]
+/**
+ * #RRGGBB or #RRGGBBAA
+ */
+export type HexColor = string
+/**
+ * CSS color keyword
+ */
+export type KeywordColor = string
+
+/**
+ * 0-1
+ */
+export type Opacity = number
+
+export interface Colors {
+  keyword: KeywordColor
+  rgb: RgbColor
+  hex: HexColor
+  hsb: HsbColor
+  hsl: HslColor
+}
+
+export type ColorType = keyof Colors
+export type ColorValue = Colors[keyof Colors]
+
+export interface ColorObject<T extends ColorType> {
+  type: T
+  value: Colors[T]
+  opacity: Opacity
+  /**
+   * Stringify the color and opacity (rgb or hex with alpha)
+   * @param withOpacity Boolean to determine if the opacity should be included in the string
+   * @returns The stringified color
+   */
+  toString: (withOpacity: boolean) => string
+}
